LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/size-container-no-principal-box-expected.txt


PASS Check that container queries is supported
FAIL (min-width: 0) does not match a container without a principal box (display:none) assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 0, 0)"
FAIL (min-width: 0) does not match a container without a principal box (display:contents) assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 0, 0)"
